backend/requirements.txt
Flask==2.3.3
google-cloud-storage==2.10.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
